+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -10,6 +10,9 @@ import java.nio.charset.StandardCharsets;
 import nextstep.jwp.exception.NotFoundException;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.handler.Handler;
+import org.apache.coyote.http11.handler.HandlerMapper;
+import org.apache.coyote.http11.http.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
