+++ tomcat/src/main/java/org/apache/coyote/http11/handler/Handler.java
@@ -1,4 +1,7 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.handler;
+
+import org.apache.coyote.http11.http.ContentType;
+import org.apache.coyote.http11.http.HttpRequest;
 
 public interface Handler {
