+++ tomcat/src/main/java/org/apache/coyote/http11/util/FileReader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.util;
 
 import java.io.File;
 import java.io.IOException;
