+++ tomcat/src/test/java/org/apache/coyote/http11/RequestTargetTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.Map;
 import java.util.stream.Stream;
+import org.apache.coyote.http11.http.RequestTarget;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
